requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
python-pptx==1.0.2
python-dotenv==1.0.1
